docs: Moves the filter feature to PhoneFilter component
