+++ src/styles/PhoneDetails.css
@@ -1,4 +1,4 @@
 .details-container {
-  margin: 8rem auto;
+  /* margin: 2rem auto 8rem auto; */
   /* padding: 30px auto; */
 }
