+++ src/styles/PhoneFilterStyles.module.css
@@ -1,0 +1,13 @@
+.label {
+  align-self: center;
+  border-radius: 10px;
+  margin: 8rem 0 2rem 0;
+}
+
+.select {
+  width: 130%;
+  height: 7vh;
+  border-radius: 10px;
+  padding: 10px;
+  font-size: 20px;
+}
