+++ src/styles/SliderStyles.module.css
@@ -3,6 +3,6 @@
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
-  margin-top: 70px;
+  margin-top: 20px;
   padding: 30px 0;
 }
